fixed spelling/wording, moved test to new file, and updated all the ReactFeatureFlag files to include new flag
