+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -38,7 +38,7 @@ export const warnAboutMissingMockScheduler = true;
 export const revertPassiveEffectsChange = false;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = false;
-export const warnAbouDefaultPropsOnFunctionComponents = false;
+export const warnAboutDefaultPropsOnFunctionComponents = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
