+++ packages/shared/forks/ReactFeatureFlags.native-oss.js
@@ -33,6 +33,7 @@ export const warnAboutMissingMockScheduler = false;
 export const revertPassiveEffectsChange = false;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = false;
+export const warnAboutDefaultPropsOnFunctionComponents = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
