+++ packages/shared/forks/ReactFeatureFlags.test-renderer.www.js
@@ -33,6 +33,7 @@ export const enableJSXTransformAPI = true;
 export const warnAboutMissingMockScheduler = true;
 export const enableUserBlockingEvents = false;
 export const enableSuspenseCallback = true;
+export const warnAboutDefaultPropsOnFunctionComponents = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
